+++ option_seealso.go
@@ -4,7 +4,10 @@ import "github.com/dogmatiq/ferrite/internal/variable"
 
 // SeeAlso is an option for a variable set that adds the variables in another
 // set, s, to the "see also" section of the generated documentation.
-func SeeAlso(s VariableSet, _ ...SeeAlsoOption) interface {
+func SeeAlso[T any](
+	s VariableSet[T],
+	_ ...SeeAlsoOption,
+) interface {
 	RequiredOption
 	OptionalOption
 	DeprecatedOption