+++ option_supersede.go
@@ -4,7 +4,10 @@ import "github.com/dogmatiq/ferrite/internal/variable"
 
 // SupersededBy is a option for a deprecated variable set that indicates the
 // variables in another set, s, should be used instead.
-func SupersededBy(s VariableSet, _ ...SupersededByOption) DeprecatedOption {
+func SupersededBy[T any](
+	s VariableSet[T],
+	_ ...SupersededByOption,
+) DeprecatedOption {
 	return option{
 		ApplyToSpecInDeprecatedSet: func(b variable.SpecBuilder) {
 			for _, v := range s.variables() {